+++ SpeedBallServer/SpeedBallServer.Test/ServerTests/JoinTests.cs
@@ -25,7 +25,7 @@ namespace SpeedBallServer.Test.ServerTests
             thirdClient = new FakeEndPoint("192.168.1.3", 5003);
 
             FakeData packet = new FakeData();
-            packet.data = new Packet((byte)PacketsCommands.Join).GetData();
+            packet.data = new Packet(PacketsCommands.Join).GetData();
             packet.endPoint = firstClient;
 
             transport.ClientEnqueue(packet);
@@ -39,7 +39,7 @@ namespace SpeedBallServer.Test.ServerTests
         public void JoinMaliciusClient()
         {
             FakeData packet = new FakeData();
-            packet.data = new Packet((byte)PacketsCommands.Join).GetData();
+            packet.data = new Packet(PacketsCommands.Join).GetData();
             packet.endPoint = maliciusClient;
 
             transport.ClientEnqueue(packet);
@@ -77,7 +77,7 @@ namespace SpeedBallServer.Test.ServerTests
         public void MalusAfterMultipleJoin()
         {
             FakeData packet = new FakeData();
-            packet.data = new Packet((byte)PacketsCommands.Join).GetData();
+            packet.data = new Packet(PacketsCommands.Join).GetData();
             packet.endPoint = firstClient;
 
             transport.ClientEnqueue(packet);
@@ -90,7 +90,7 @@ namespace SpeedBallServer.Test.ServerTests
         public void ClientsAmountAfterMultipleJoins()
         {
             FakeData packet = new FakeData();
-            packet.data = new Packet((byte)PacketsCommands.Join).GetData();
+            packet.data = new Packet(PacketsCommands.Join).GetData();
             packet.endPoint = secondClient;
             transport.ClientEnqueue(packet);
             packet.endPoint = thirdClient;
